+++ lib/core/routes/app_router.dart
@@ -10,8 +10,6 @@ import 'package:flower_app/features/home/domain/entities/product_entity.dart';
 import 'package:flower_app/features/home/presentation/pages/home_screen.dart';
 import 'package:flower_app/features/nav/nav_bar.dart';
 import 'package:flower_app/features/occasion/presentation/pages/occasion_page.dart';
-import 'package:flower_app/features/profile/presentation/cubit/profile_cubit.dart';
-import 'package:flower_app/features/profile/presentation/pages/add_adress_screen.dart';
 import 'package:flower_app/features/profile/presentation/pages/profile_reset_password.dart';
 import 'package:flower_app/features/splash/splash_screen.dart';
 import 'package:flower_app/core/common_widgets/product_details_page/product_details_page.dart';
@@ -19,6 +17,8 @@ import 'package:flower_app/core/common_widgets/product_details_page/product_deta
 import 'package:flower_app/features/auth/presentation/pages/reset_password_page.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
+import '../../features/add_address/presentation/manager/add_address_cubit.dart';
+import '../../features/add_address/presentation/pages/add_adress_screen.dart';
 import '../../features/profile/domain/entities/user_data.dart';
 import '../../features/profile/presentation/pages/edit_profile_page.dart';
 
@@ -118,7 +118,7 @@ Route<dynamic>? generateRoute(RouteSettings settings) {
       return MaterialPageRoute(
         settings: settings,
         builder: (_) => BlocProvider(
-          create: (context) => getIt<ProfileCubit>(),
+          create: (context) => getIt<AddAddressCubit>(),
           child: const AddAddressScreen()),
       );
