+++ lib/features/add_address/data/data_sources/add_address_remote_data_source.dart
@@ -1,0 +1,8 @@
+
+
+
+import '../models/add_adress_model/add_adress_request.dart';
+
+abstract class AddAddressRemoteDataSource {
+  Future<void> addAddress(AddAdressRequest addAdressRequest);
+}
