+++ lib/features/add_address/data/data_sources/add_address_remote_data_source_impl.dart
@@ -1,0 +1,20 @@
+import 'package:injectable/injectable.dart';
+
+import '../../../../core/app_data/api/api_client.dart';
+import '../../../../core/app_data/api/api_constants.dart';
+import '../models/add_adress_model/add_adress_request.dart';
+import 'add_address_remote_data_source.dart';
+@Injectable(as: AddAddressRemoteDataSource)
+class AddAddressRemoteDataSourceImpl implements AddAddressRemoteDataSource{
+  final ApiClient _apiClient;
+
+  AddAddressRemoteDataSourceImpl(this._apiClient);
+  @override
+  Future<void> addAddress(AddAdressRequest addAdressRequest) async {
+    await _apiClient.patch(
+      ApiConstants.addressesEndPoint,
+      data: addAdressRequest.toJson(),
+      requiresToken: true,
+    );
+  }
+}
