+++ lib/features/add_address/domain/use_cases/add_adress_use_case.dart
@@ -1,0 +1,12 @@
+import 'package:either_dart/either.dart';
+import 'package:injectable/injectable.dart';
+
+import '../../data/models/add_adress_model/add_adress_request.dart';
+import '../repositories/add_address_repo.dart';
+@injectable
+class AddAddressUseCase {
+  final AddAddressRepo _addAddressRepo;
+  AddAddressUseCase(this._addAddressRepo);
+
+  Future<Either<Exception, void>> call(AddAdressRequest addAdressRequset) async => await _addAddressRepo.addAddress(addAdressRequset);
+}
